+++ src/Web/Controllers/ClientController.cs
@@ -28,60 +28,94 @@ namespace Web.Controllers
         [HttpGet]
         public IActionResult GetAll()
         {
-            return Ok(_service.GetAllClients());
+            // Obtener el claim de rol, si existe
+            var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role);
+
+            // Verificar si el claim existe y su valor es "Admin"
+            if (roleClaim != null && roleClaim.Value == "Admin")
+            {
+                return Ok(_service.GetAllClients());
+            }
+            // Si el rol no es Admin o el claim no existe, prohibir acceso
+            return Forbid();
         }
 
         [HttpGet("{id}")]
         public IActionResult GetById([FromRoute] int id)
         {
-            var client = _service.Get(id);
-            if (client == null)
+            var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role);
+            if (roleClaim.Value == "Admin")
             {
-                return NotFound($"No se encontró ningún cliente con el ID: {id}");
+                var client = _service.Get(id);
+                if (client == null)
+                {
+                    return NotFound($"No se encontró ningún cliente con el ID: {id}");
+                }
+                return Ok(client);
             }
-            return Ok(client);
+            return Forbid();
         }
 
         [HttpGet("{name}")]
         public IActionResult GetByName([FromRoute] string name)
         {
-            var client = _service.Get(name);
-            if (client == null)
+            var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role);
+            if (roleClaim.Value == "Admin")
             {
-                return NotFound($"No se encontró ningún cliente con el nombre: {name}");
+                var client = _service.Get(name);
+                if (client == null)
+                {
+                    return NotFound($"No se encontró ningún cliente con el nombre: {name}");
+                }
+                return Ok(client);
             }
-            return Ok(client);
+            return Forbid();
         }
 
         [HttpPost]
         public IActionResult Add([FromBody] ClientCreateRequest body)
         {
-            var newClient = _service.AddClient(body);
-            return Ok($"Creado el Cliente con el ID: {newClient}");
+            var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role);
+            if (roleClaim.Value == "Admin" || roleClaim.Value == "Client")
+            {
+                var newClient = _service.AddClient(body);
+                return Ok($"Creado el Cliente con el ID: {newClient}");
+            }
+            return Forbid();
         }
 
         [HttpDelete("{id}")]
         public IActionResult DeleteClient([FromRoute] int id)
         {
-            var existingClient = _service.Get(id);
-            if (existingClient == null)
+            var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role);
+            if (roleClaim.Value == "Admin")
             {
-                return NotFound($"No se encontró ningún Cliente con el ID: {id}");
+                var existingClient = _service.Get(id);
+                if (existingClient == null)
+                {
+                    return NotFound($"No se encontró ningún Cliente con el ID: {id}");
+                }
+                _service.DeleteClient(id);
+                return Ok($"Cliente con ID: {id} eliminado");
             }
-            _service.DeleteClient(id);
-            return Ok($"Cliente con ID: {id} eliminado");                
+            return Forbid();               
         }
 
         [HttpPut("{id}")]
         public IActionResult UpdateClient([FromRoute] int id, [FromBody] ClientUpdateRequest request)
         {
-            var existingClient = _service.Get(id);
-            if (existingClient == null)
+            var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role);
+            if (roleClaim.Value == "Admin" || roleClaim.Value == "Client")
             {
-                return NotFound($"No se encontró ningún Cliente con el ID: {id}");
+                var existingClient = _service.Get(id);
+                if (existingClient == null)
+                {
+                    return NotFound($"No se encontró ningún Cliente con el ID: {id}");
+                }
+                _service.UpdateClient(id, request);
+                return Ok($"Cliente con ID: {id} actualizado correctamente");
             }
-            _service.UpdateClient(id, request);
-            return Ok($"Cliente con ID: {id} actualizado correctamente");
+            return Forbid();
         }
     }
 }
